Add 추천 service 로직 {스부 > 장고 > ES} 로 변경 [점수제 변경] hover : 1 click, shopList : 2 order: 3
hover가 생각보다 아주 양이 많아!
그래서 최근 50개 행동 가져와서 : ubic_click_action
{카테고리 : 점수} 의 배열을 장고에서 계산해서
response 할거야

+++ src/main/java/com/ubic/shop/service/ShopListService.java
@@ -58,7 +58,11 @@ public class ShopListService {
     @Transactional
     public void cancelShopList(Long shopListId) {
         //취소
-        shopListRepository.deleteById(shopListId);
+        if(shopListRepository.findById(shopListId).isPresent()){
+            shopListRepository.deleteById(shopListId);
+        }
+        em.flush();
+        em.clear();
     }
 
     /*장바구니 수정*/
